+++ step1/src/App.ts
@@ -1,4 +1,4 @@
-import {Component} from "~_core/Component";
+import {Component} from "~_core";
 import {Header} from "~components/Header";
 import {YoutubeSearchModal} from "~components/YoutubeSearchModal";
 import {VideoClip, VideoClipType} from "~components/VideoClip";
