+++ step1/src/_core/index.ts
@@ -1,0 +1,5 @@
+export * from './Component';
+export * from './Store';
+export * from './Observer';
+export * from './Repository';
+export * from './RestClient';
