+++ step1/src/clients/youtubeRestClient.ts
@@ -1,13 +1,13 @@
-import {RestClient} from "~_core/RestClient";
+import {RestClient} from "~_core";
 import {YOUTUBE_API_KEY} from "~clients/secret";
 
 export class YoutubeRestClient extends RestClient {
   constructor(private readonly apiKey: string) {
-    super(`https://www.googleapis.com/youtube/v3`);
+    super(`https://www.googleapis.com/youtube/v3/`);
   }
 
   public get <T>(uri: string): Promise<T> {
-    const appendedUri = new URL(uri);
+    const appendedUri = new URL(`.${uri}`, this.baseUrl);
     appendedUri.searchParams.set('key', this.apiKey);
     return super.get<T>(appendedUri.toString());
   }
