+++ step1/src/components/RecentSearches.ts
@@ -1,4 +1,4 @@
-import {Component} from "~_core/Component";
+import {Component} from "~_core";
 
 interface Props {
   items: string[];
