+++ step1/src/domain/Youtube.ts
@@ -1,34 +1,13 @@
-export interface YoutubeClient {
-  init: (option: { apiKey: string }) => Promise<void>,
-  youtube: {
-    search: {
-      list: (option: YoutubeSearchRequestOptions) => Promise<YoutubeSearchResult>
-    }
-  }
-}
-
-/**
- * @link {https://developers.google.com/youtube/v3/docs/search/list?hl=ko}
- */
-export interface YoutubeSearchRequestOptions {
-  q: string;
-  part: string;
-  order: string;
-  maxResults: number;
-}
-
 export interface YoutubeSearchResult {
-  result: {
-    nextPageToken?: string;
-    prevPageToken?: string;
-    etag: string;
-    kind: string;
-    pageInfo: {
-      totalResults: number;
-      resultsPerPage: 10
-    }
-    items: YoutubeClipItem[];
+  nextPageToken?: string;
+  prevPageToken?: string;
+  etag: string;
+  kind: string;
+  pageInfo: {
+    totalResults: number;
+    resultsPerPage: 10
   }
+  items: YoutubeClipItem[];
 }
 
 export interface YoutubeClipItem {
