+++ step1/src/services/index.ts
@@ -1,1 +1,2 @@
 export * from './youtubeService';
+export * from './recentSearchesService';
